@@ -15,9 +15,12 @@
     "postcss": "^8.4.27",
     "postcss-nesting": "^12.0.0",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^9.1.1",
+    "react-router-dom": "^6.23.0",
     "react-scripts": "^5.0.0",
+    "redux": "^5.0.1",
     "styled-components": "^5.3.0",
     "swiper": "^10.1.0",
     "web-vitals": "^1.0.1"
@@ -50,6 +53,8 @@
     ]
   },
   "devDependencies": {
+    "@types/react": "^18.3.0",
+    "@types/react-dom": "^18.3.0",
     "eslint": "^7.32.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-plugin-import": "^2.24.2",
