@@ -1,13 +1,9 @@
-import React, { useState } from "react";
+import React, { createContext, useState } from "react";
+import { Route, BrowserRouter as Router, Routes } from "react-router-dom";
 import "./App.module.css";
 import styles from "./App.module.css";
-import Header from "./components/Header";
+import { HomePage } from "./Pages/Home";
 import Navbar from "./components/Navbar";
-// import Gallery from "./components/Gallery";
-import { createContext } from "react";
-import Projects from "./components/Projects";
-import Resume from "./components/Resume";
-
 export const ThemeContext = createContext(null);
 
 function App() {
@@ -17,13 +13,12 @@ function App() {
 
   return (
     <ThemeContext.Provider value={{ theme, setTheme }}>
-      <div className={className}>
-        <Navbar />
-        <Header />
-        <Resume className={styles.section} />
-        {/* <Gallery /> */}
-        <Projects className={styles.section} />
-      </div>
+      <Navbar />
+      <Router>
+        <Routes>
+          <Route path="/" element={<HomePage />} />
+        </Routes>
+      </Router>
     </ThemeContext.Provider>
   );
 }
